Break up character commands
